Remove atmos holdovers while at it!
