+++ Content.Server/Administration/Commands/SetAdminOOC.cs
@@ -1,6 +1,5 @@
 
 using Content.Server.Database;
-using Content.Server.Interfaces;
 using Content.Server.Preferences.Managers;
 using Content.Shared.Administration;
 using Robust.Server.Player;
