+++ Content.Server/Administration/Commands/SetOutfitCommand.cs
@@ -1,6 +1,5 @@
 using Content.Server.Administration.UI;
 using Content.Server.EUI;
-using Content.Server.Interfaces;
 using Content.Server.Inventory.Components;
 using Content.Server.Items;
 using Content.Server.PDA;
