+++ Content.Server/Atmos/Commands/AddGasCommand.cs
@@ -1,38 +1,31 @@
 ﻿using Content.Server.Administration;
-using Content.Server.Atmos.Components;
 using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Administration;
 using Content.Shared.Atmos;
 using Robust.Shared.Console;
 using Robust.Shared.GameObjects;
-using Robust.Shared.IoC;
 using Robust.Shared.Map;
 using Robust.Shared.Maths;
 
-namespace Content.Server.Commands.Atmos
+namespace Content.Server.Atmos.Commands
 {
     [AdminCommand(AdminFlags.Debug)]
-    public class SetTemperatureCommand : IConsoleCommand
+    public class AddGasCommand : IConsoleCommand
     {
-        public string Command => "settemp";
-        public string Description => "Sets a tile's temperature (in kelvin).";
-        public string Help => "Usage: settemp <X> <Y> <GridId> <Temperature>";
+        public string Command => "addgas";
+        public string Description => "Adds gas at a certain position.";
+        public string Help => "addgas <X> <Y> <GridId> <Gas> <moles>";
 
         public void Execute(IConsoleShell shell, string argStr, string[] args)
         {
-            if (args.Length < 4) return;
+            if (args.Length < 5) return;
             if(!int.TryParse(args[0], out var x)
                || !int.TryParse(args[1], out var y)
                || !int.TryParse(args[2], out var id)
-               || !float.TryParse(args[3], out var temperature)) return;
+               || !(AtmosCommandUtils.TryParseGasID(args[3], out var gasId))
+               || !float.TryParse(args[4], out var moles)) return;
 
             var gridId = new GridId(id);
-
-            if (temperature < Atmospherics.TCMB)
-            {
-                shell.WriteLine("Invalid temperature.");
-                return;
-            }
 
             var atmosphereSystem = EntitySystem.Get<AtmosphereSystem>();
             var indices = new Vector2i(x, y);
@@ -44,7 +37,7 @@ namespace Content.Server.Commands.Atmos
                 return;
             }
 
-            tile.Temperature = temperature;
+            tile.AdjustMoles(gasId, moles);
         }
     }
 }
