+++ Content.Server/Atmos/Commands/AddUnsimulatedAtmosCommand.cs
@@ -6,7 +6,7 @@ using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Map;
 
-namespace Content.Server.Commands.Atmos
+namespace Content.Server.Atmos.Commands
 {
     [AdminCommand(AdminFlags.Debug)]
     public class AddUnsimulatedAtmosCommand : IConsoleCommand
