+++ Content.Server/Atmos/Commands/RemoveGasCommand.cs
@@ -1,22 +1,19 @@
 ﻿using Content.Server.Administration;
-using Content.Server.Atmos.Components;
 using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Administration;
-using Content.Shared.Atmos;
 using Robust.Shared.Console;
 using Robust.Shared.GameObjects;
-using Robust.Shared.IoC;
 using Robust.Shared.Map;
 using Robust.Shared.Maths;
 
-namespace Content.Server.Commands.Atmos
+namespace Content.Server.Atmos.Commands
 {
     [AdminCommand(AdminFlags.Debug)]
-    public class AddGasCommand : IConsoleCommand
+    public class RemoveGasCommand : IConsoleCommand
     {
-        public string Command => "addgas";
-        public string Description => "Adds gas at a certain position.";
-        public string Help => "addgas <X> <Y> <GridId> <Gas> <moles>";
+        public string Command => "removegas";
+        public string Description => "Removes an amount of gases.";
+        public string Help => "removegas <X> <Y> <GridId> <amount> <ratio>\nIf <ratio> is true, amount will be treated as the ratio of gas to be removed.";
 
         public void Execute(IConsoleShell shell, string argStr, string[] args)
         {
@@ -24,8 +21,8 @@ namespace Content.Server.Commands.Atmos
             if(!int.TryParse(args[0], out var x)
                || !int.TryParse(args[1], out var y)
                || !int.TryParse(args[2], out var id)
-               || !(AtmosCommandUtils.TryParseGasID(args[3], out var gasId))
-               || !float.TryParse(args[4], out var moles)) return;
+               || !float.TryParse(args[3], out var amount)
+               || !bool.TryParse(args[4], out var ratio)) return;
 
             var gridId = new GridId(id);
 
@@ -39,7 +36,11 @@ namespace Content.Server.Commands.Atmos
                 return;
             }
 
-            tile.AdjustMoles(gasId, moles);
+            if (ratio)
+                tile.RemoveRatio(amount);
+            else
+                tile.Remove(amount);
         }
     }
+
 }
