+++ Content.Server/Atmos/Components/GasMixtureHolderComponent.cs
@@ -1,5 +1,4 @@
-﻿using Content.Server.Interfaces;
-using Robust.Shared.GameObjects;
+﻿using Robust.Shared.GameObjects;
 using Robust.Shared.Serialization.Manager.Attributes;
 using Robust.Shared.ViewVariables;
