+++ Content.Server/Atmos/IGasMixtureHolder.cs
@@ -1,8 +1,7 @@
-﻿using Content.Server.Atmos;
-using Content.Server.Atmos.EntitySystems;
+﻿using Content.Server.Atmos.EntitySystems;
 using Robust.Shared.GameObjects;
 
-namespace Content.Server.Interfaces
+namespace Content.Server.Atmos
 {
     public interface IGasMixtureHolder
     {
