+++ Content.Server/Atmos/Reactions/PlasmaFireReaction.cs
@@ -1,7 +1,6 @@
 using System;
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Coordinates.Helpers;
-using Content.Server.Interfaces;
 using Content.Shared.Atmos;
 using JetBrains.Annotations;
 using Robust.Server.GameObjects;
