+++ Content.Server/Atmos/Reactions/TritiumFireReaction.cs
@@ -1,6 +1,5 @@
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Coordinates.Helpers;
-using Content.Server.Interfaces;
 using Content.Shared.Atmos;
 using JetBrains.Annotations;
 using Robust.Server.GameObjects;
