+++ Content.Server/Atmos/Reactions/WaterVaporReaction.cs
@@ -1,6 +1,5 @@
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Fluids.Components;
-using Content.Server.Interfaces;
 using Content.Shared.Chemistry.Reagent;
 using Content.Shared.Chemistry.Solution;
 using Content.Shared.Maps;
