+++ Content.Server/Disposal/Unit/Components/DisposalHolderComponent.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Disposal.Tube.Components;
-using Content.Server.Interfaces;
 using Content.Server.Items;
 using Content.Shared.Atmos;
 using Content.Shared.Body.Components;
