+++ Content.Server/NodeContainer/NodeGroups/PipeNet.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Content.Server.Atmos;
 using Content.Server.Atmos.Components;
 using Content.Server.Atmos.EntitySystems;
-using Content.Server.Interfaces;
 using Content.Server.NodeContainer.Nodes;
 using Content.Shared.Atmos;
 using Robust.Shared.GameObjects;
