+++ Content.Server/NodeContainer/Nodes/PipeNode.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
-using Content.Server.Interfaces;
 using Content.Server.NodeContainer.EntitySystems;
 using Content.Server.NodeContainer.NodeGroups;
 using Content.Shared.Atmos;
