+++ Resources/Locale/en-US/atmos/firelock-component.ftl
