+++ Resources/Locale/en-US/atmos/flammable-component.ftl
