+++ Resources/Locale/en-US/atmos/gas-analyzer-component.ftl
