+++ Resources/Locale/en-US/atmos/gas-canister-component.ftl
