+++ Resources/Locale/en-US/atmos/gas-tank-component.ftl
